refactor: split RuntimePage into orchestrator + 11 section components
Extract 1026-line monolithic RuntimePage into a lean orchestrator (~45
lines) that owns WebSocket hooks and delegates rendering to dedicated
section components in pages/runtime/.  Extract shared CollapsibleSection
component and typed interfaces into types/runtime.ts (20 exports).

Constraint: must preserve identical behaviour — no feature changes
Confidence: high
Scope-risk: moderate — largest single-file refactor to date
